Add files via upload
Charged particle radiography is a diagnostic technique used to study high-energy density plasmas by examining the behavior of charged particles, typically protons, as they interact with the electric and magnetic fields within the plasma. This technique involves a bright source of charged particles, which illuminates the plasma region of interest, and a detector plane to capture the patterns created by the deflection of these particles due to the plasma's electric and magnetic fields. The information obtained from these radiographs can provide insights into the behavior of these fields within the plasma.